Moved client into client directory
